Rename isCustomProperty
