+++ src/rules/property-blacklist/index.js
@@ -3,7 +3,7 @@ import { isString } from "lodash"
 import {
   isStandardDeclaration,
   matchesStringOrRegExp,
-  propertyIsCustom,
+  isCustomProperty,
   report,
   ruleMessages,
   validateOptions,
@@ -28,7 +28,7 @@ export default function (blacklistInput) {
       if (!isStandardDeclaration(decl)) { return }
 
       const { prop } = decl
-      if (propertyIsCustom(prop)) { return }
+      if (isCustomProperty(prop)) { return }
       if (!matchesStringOrRegExp(vendor.unprefixed(prop), blacklist)) { return }
 
       report({
