+++ src/utils/index.js
@@ -25,7 +25,7 @@ export { default as matchesStringOrRegExp } from "./matchesStringOrRegExp"
 export { default as mediaQueryParamIndexOffset } from "./mediaQueryParamIndexOffset"
 export { default as optionsHaveException } from "./optionsHaveException"
 export { default as optionsHaveIgnored } from "./optionsHaveIgnored"
-export { default as propertyIsCustom } from "./propertyIsCustom"
+export { default as isCustomProperty } from "./isCustomProperty"
 export { default as rawNodeString } from "./rawNodeString"
 export { default as report } from "./report"
 export { default as ruleMessages } from "./ruleMessages"
